+++ android/app/src/main/java/com/nosiah/discipline/MainActivity.java
@@ -1,0 +1,5 @@
+package com.nosiah.discipline;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}